+++ cmd/cmd_debug.go
@@ -10,6 +10,15 @@ func addDebugOption(cmd *cobra.Command, usage string) {
 	cmd.Flags().BoolP("debug", "", false, "Debug mode: "+usage)
 }
 
+// isInDebugMode returns true if the --debug flag is set
+func isInDebugMode(cmd *cobra.Command) bool {
+	debug, err := cmd.Flags().GetBool("debug")
+	if err != nil {
+		return false
+	}
+	return debug
+}
+
 // debugPrintf prints the format string and arguments if the --debug flag is set
 func debugPrintf(cmd *cobra.Command, format string, a ...interface{}) {
 	if isInDebugMode(cmd) {
@@ -22,12 +31,4 @@ func debugPrintln(cmd *cobra.Command, a ...interface{}) {
 	if isInDebugMode(cmd) {
 		fmt.Println(a...)
 	}
-}
-
-func isInDebugMode(cmd *cobra.Command) bool {
-	debug, err := cmd.Flags().GetBool("debug")
-	if err != nil {
-		return false
-	}
-	return debug
 }