+++ cmd/root.go
@@ -11,7 +11,7 @@ import (
 // rootCmd represents the base command when called without any subcommands
 var rootCmd = &cobra.Command{
 	Use:   "twigtango",
-	Short: "Parsw twig-line files",
+	Short: "Parse twig-line files",
 	Long:  ``,
 	// Uncomment the following line if your bare application
 	// has an action associated with it: