+++ tango/dotenv.go
@@ -5,10 +5,11 @@ import (
 	"os"
 )
 
-func LoadDotEnv(path string) (err error) {
+// LoadDotEnv loads the .env file in the given directory
+func LoadDotEnv(dir string) (err error) {
 	// check if .env file exists
-	if _, err = os.Stat(path + "/.env"); os.IsNotExist(err) {
+	if _, err = os.Stat(dir + "/.env"); os.IsNotExist(err) {
 		return nil
 	}
-	return godotenv.Load(path + "/.env")
+	return godotenv.Load(dir + "/.env")
 }